+++ supervised_benchmarks/metric_protocols.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from abc import abstractmethod
 from typing import Literal, Protocol, NamedTuple, List, TypeVar, Tuple, Any, Callable
 
-from supervised_benchmarks.dataset_protocols import DataArray
+from numpy.typing import NDArray
 from variable_protocols.variables import Variable
 
 PairMetricType = Literal['mean_acc', 'categorical_acc']
@@ -40,7 +40,7 @@ class MetricResult(NamedTuple):
     result_type: PairMetricType
 
 
-Measure = Callable[[DataArray, DataArray], MetricResult]
+Measure = Callable[[NDArray, NDArray], MetricResult]
 
 
 class PairMetricImp(NamedTuple):