+++ supervised_benchmarks/tests/test_uci_income.py
@@ -6,7 +6,9 @@ from catboost import CatBoostClassifier
 
 from numpy.typing import NDArray
 
+from supervised_benchmarks.benchmark import BenchmarkConfig
 from supervised_benchmarks.dataset_protocols import FixedTrain, FixedTest, DataUnit
+from supervised_benchmarks.metrics import get_pair_metric
 from supervised_benchmarks.ports import Port
 from supervised_benchmarks.protocols import Performer
 from supervised_benchmarks.sampler import FixedEpochSamplerConfig, FullBatchSamplerConfig
@@ -31,7 +33,7 @@ class BoostModelConfig(NamedTuple):
         print(tr.content_map[AnyNetDiscreteOut])
         clf.fit(tr.content_map[AnyNetDiscrete], tr.content_map[AnyNetDiscreteOut])
 
-        return BoostPerformer(classifier=clf, repertoire={AnyNetDiscreteOut})
+        return BoostPerformer(classifier=clf, repertoire=frozenset({AnyNetDiscreteOut}))
 
 
 class BoostPerformer(NamedTuple):
@@ -85,9 +87,9 @@ def test_boost_init():
              AnyNetDiscreteOut: uci_income_out_anynet_discrete}
     data_config = UciIncomeDataConfig(base_path=Path('/Data/uci'), query=query)
     data_pool = data_config.get_data()
-    tr = data_pool.fixed_subsets[FixedTrain]
+    tst = data_pool.fixed_subsets[FixedTest]
     sampler_config = FixedEpochSamplerConfig(512)
-    sampler = sampler_config.get_sampler(tr)
+    sampler = sampler_config.get_sampler(tst)
 
     mini_batch = next(sampler.iter)
     config = BoostModelConfig(
@@ -95,3 +97,17 @@ def test_boost_init():
     performer = config.prepare()
     result = performer.perform(data_src=mini_batch, tgt=AnyNetDiscreteOut)
     print((result == mini_batch[AnyNetDiscreteOut]).mean())
+
+
+def test_benchmark():
+    query = {AnyNetDiscrete: uci_income_in_anynet_discrete,
+             AnyNetDiscreteOut: uci_income_out_anynet_discrete}
+    data_config = UciIncomeDataConfig(base_path=Path('/Data/uci'), query=query)
+    benchmark_config = BenchmarkConfig(
+        metrics={AnyNetDiscreteOut: get_pair_metric('mean_acc', data_config.query[AnyNetDiscreteOut])},
+        on=FixedTest)
+    model_config = BoostModelConfig(
+        ports={AnyNetDiscreteOut: uci_income_out_anynet_discrete, AnyNetDiscrete: uci_income_in_anynet_discrete})
+    performer = model_config.prepare()
+    z = benchmark_config.bench(data_config, performer)
+    print(z)